pyproject.toml
[tool.poetry]
name = "taiwanreservoir"
version = "0.1.0"
description = ""
authors = ["tasi788 <tdc@tdccc.com.tw>"]

[tool.poetry.dependencies]
python = "^3.8.6"
requests = "^2.25.1"
pyquery = "^1.4.3"
coloredlogs = "^15.0"
dataclasses-json = "^0.5.4"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
